+++ src/lang/code.ts
@@ -1,0 +1,30 @@
+/**
+ * @category Localization
+ */
+export type LangCode =
+	| 'ar_ae'
+	| 'cs_cz'
+	| 'da_dk'
+	| 'de_de'
+	| 'en_gb'
+	| 'en_us'
+	| 'es_es'
+	| 'fi_fi'
+	| 'fr_fr'
+	| 'in_id'
+	| 'it_it'
+	| 'ja_jp'
+	| 'ko_kr'
+	| 'ms_my'
+	| 'nl_nl'
+	| 'no_no'
+	| 'pl_pl'
+	| 'pt_br'
+	| 'ro_ro'
+	| 'ru_ru'
+	| 'sv_se'
+	| 'th_th'
+	| 'tl_ph'
+	| 'tr_tr'
+	| 'zh_cn'
+	| 'zh_tw';