+++ src/lifecycle/entity/phase.ts
@@ -23,36 +23,35 @@
  *
  */
 
-import type {LifecycleListener} from '../../../lifecycle/listener';
-import type {LifecycleNetworkCnx} from '../cnx';
-
 /**
- * Delegate interface for classes implementing Network Client listeners.
+ * Identifiers for each phase in the Entity Lifecycle flow.
  *
- * @category Lifecycle
+ * @category Lifecycle - Entities
  */
-export interface LifecycleNetworkCnxDelegate {
-	lifecycle: LifecycleNetworkCnx;
-	cnxDidClose?: LifecycleListener;
-	cnxDidConnect?: LifecycleListener;
-	cnxDidFailConnect?: LifecycleListener;
-	cnxDidFailHandshake?: LifecycleListener;
-	cnxDidHandshake?: LifecycleListener;
-	cnxDidInit?: LifecycleListener;
-	cnxDidLoad?: LifecycleListener;
-	cnxDidOpen?: LifecycleListener;
-	cnxDidRcvMsg?: LifecycleListener;
-	cnxDidReconnect?: LifecycleListener;
-	cnxDidSndMsg?: LifecycleListener;
-	cnxDidStartConnect?: LifecycleListener;
-	cnxDidStopConnect?: LifecycleListener;
-	cnxOnPing?: LifecycleListener;
-	cnxOnPong?: LifecycleListener;
-	cnxWillHandshake?: LifecycleListener;
-	cnxWillInit?: LifecycleListener;
-	cnxWillLoad?: LifecycleListener;
-	cnxWillOpen?: LifecycleListener;
-	cnxWillReconnect?: LifecycleListener;
-	cnxWillStartConnect?: LifecycleListener;
-	cnxWillStopConnect?: LifecycleListener;
-}
+export type LifecycleEntityPhase =
+	| 'entityWillInit'
+	| 'entityDidInit'
+	| 'sceneWillLoad'
+	| 'sceneDidLoad'
+	| 'entityWillLoad'
+	| 'entityDidLoad'
+	| 'entityWillSpawn'
+	| 'entityDidSpawn'
+	| 'entityWillAppear'
+	| 'entityDidAppear'
+	| 'sceneWillStart'
+	| 'sceneDidStart'
+	| 'entityWillStart'
+	| 'entityDidStart'
+	| 'sceneWillPause'
+	| 'entityWillPause'
+	| 'sceneDidUnpause'
+	| 'entityDidUnpause'
+	| 'entityWillHide'
+	| 'entityDidHide'
+	| 'entityWillDespawn'
+	| 'entityDidDespawn'
+	| 'sceneWillStop'
+	| 'entityWillStop'
+	| 'entityDidStop'
+	| 'memoryWarning';